multipage/requirements.txt
streamlit-option-menu==0.3.2
googlesearch-python==1.1.0
beautifulsoup4==4.9.3
requests[security]==2.25.1